requirements.txt
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
tabulate>=0.9.0
colorama>=0.4.6
tqdm>=4.66.0
geopy>=2.4.0
census>=0.8.22

